# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SLM Fine-Tuning Platform",
    "description": "Backend-only PoC for an automated Small Language Model fine-tuning platform. Generates synthetic data via OpenRouter, fine-tunes ≤3B models with Unsloth + QLoRA, tracks every run in MLflow, and serves results via Ollama.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/trainings": {
      "post": {
        "tags": [
          "trainings"
        ],
        "summary": "Start a training job (manual or HPO)",
        "operationId": "start_training_api_v1_trainings_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ManualTrainingRequest"
                  },
                  {
                    "$ref": "#/components/schemas/HPOTrainingRequest"
                  }
                ],
                "discriminator": {
                  "propertyName": "mode",
                  "mapping": {
                    "manual": "#/components/schemas/ManualTrainingRequest",
                    "hpo": "#/components/schemas/HPOTrainingRequest"
                  }
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobAcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "trainings"
        ],
        "summary": "List training jobs (filter by project / status)",
        "operationId": "list_trainings_api_v1_trainings_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TrainingResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trainings/{training_id}": {
      "get": {
        "tags": [
          "trainings"
        ],
        "summary": "Get a training job by id",
        "operationId": "get_training_api_v1_trainings__training_id__get",
        "parameters": [
          {
            "name": "training_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Training Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "trainings"
        ],
        "summary": "Hard-delete a finished training job (and its model artifact, if any)",
        "operationId": "delete_training_api_v1_trainings__training_id__delete",
        "parameters": [
          {
            "name": "training_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Training Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trainings/{training_id}/cancel": {
      "post": {
        "tags": [
          "trainings"
        ],
        "summary": "Cancel a running / pending training job",
        "description": "Cancel an in-flight training job (idempotent on an already-terminal one).\n\n**BREAKING CHANGE note**: this used to be documented as a POST alias for\n`DELETE /{training_id}`, but `DELETE` was repurposed into a hard-delete\n(see that route above) — this is now the *only* cancel entry point for\ntrainings. Callers that used to `DELETE` to cancel (e.g.\n`smart-model-tune`'s `engineApi.ts`) must switch to this endpoint.",
        "operationId": "cancel_training_post_api_v1_trainings__training_id__cancel_post",
        "parameters": [
          {
            "name": "training_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Training Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Cancel Training Post Api V1 Trainings  Training Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trainings/{training_id}/mlflow-url": {
      "get": {
        "tags": [
          "trainings"
        ],
        "summary": "Resolve the MLflow run URL for a training job",
        "operationId": "get_mlflow_url_api_v1_trainings__training_id__mlflow_url_get",
        "parameters": [
          {
            "name": "training_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Training Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MlflowUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/trainings/{training_id}/loss-history": {
      "get": {
        "tags": [
          "trainings"
        ],
        "summary": "Lightweight train_loss + eval_loss series for chart components",
        "operationId": "get_training_loss_history_api_v1_trainings__training_id__loss_history_get",
        "parameters": [
          {
            "name": "training_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Training Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingLossHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HPOTrainingRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "base_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Model",
            "description": "Override DEFAULT_BASE_MODEL (e.g. 'unsloth/Llama-3.2-1B-Instruct-bnb-4bit')."
          },
          "training_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Name",
            "description": "MLflow run name; defaults to project + timestamp. When provided, must be ollama-tag-safe: lowercase letters, digits, '.', '_', '-' only, starting with a letter or digit, 1-63 chars total. Must be unique within the owning project's owner scope."
          },
          "auto_export": {
            "type": "boolean",
            "title": "Auto Export",
            "description": "If true, automatically kick off GGUF export when this training job completes successfully.",
            "default": false
          },
          "auto_evaluate": {
            "type": "boolean",
            "title": "Auto Evaluate",
            "description": "If true, automatically kick off evaluation once the auto-export step (or training itself, if auto_export is false) completes. Has no effect unless the export it depends on succeeds.",
            "default": false
          },
          "mode": {
            "type": "string",
            "const": "hpo",
            "title": "Mode",
            "default": "hpo"
          },
          "hpo_config": {
            "$ref": "#/components/schemas/HPOConfig"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "dataset_id",
          "hpo_config"
        ],
        "title": "HPOTrainingRequest",
        "examples": [
          {
            "dataset_id": "00000000-0000-0000-0000-000000000010",
            "hpo_config": {
              "direction": "minimize",
              "n_trials": 6,
              "objective_metric": "eval_loss",
              "pruner": "median",
              "sampler": "tpe",
              "search_space": {
                "gradient_accumulation_steps": {
                  "choices": [
                    4,
                    8,
                    16
                  ],
                  "type": "categorical"
                },
                "learning_rate": {
                  "high": 0.0005,
                  "log": true,
                  "low": 1e-05,
                  "type": "float"
                },
                "lora_alpha": {
                  "choices": [
                    16,
                    32,
                    64
                  ],
                  "type": "categorical"
                },
                "lora_r": {
                  "choices": [
                    8,
                    16,
                    32
                  ],
                  "type": "categorical"
                },
                "num_train_epochs": {
                  "high": 4,
                  "low": 2,
                  "type": "int"
                }
              },
              "timeout_seconds": 14400
            },
            "mode": "hpo",
            "project_id": "00000000-0000-0000-0000-000000000001",
            "training_name": "qa-policy-search"
          }
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "description": "Lifecycle of an async Celery-backed job (SDG / training / evaluation)."
      },
...
      "ManualTrainingRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "base_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Model",
            "description": "Override DEFAULT_BASE_MODEL (e.g. 'unsloth/Llama-3.2-1B-Instruct-bnb-4bit')."
          },
          "training_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Name",
            "description": "MLflow run name; defaults to project + timestamp. When provided, must be ollama-tag-safe: lowercase letters, digits, '.', '_', '-' only, starting with a letter or digit, 1-63 chars total. Must be unique within the owning project's owner scope."
          },
          "auto_export": {
            "type": "boolean",
            "title": "Auto Export",
            "description": "If true, automatically kick off GGUF export when this training job completes successfully.",
            "default": false
          },
          "auto_evaluate": {
            "type": "boolean",
            "title": "Auto Evaluate",
            "description": "If true, automatically kick off evaluation once the auto-export step (or training itself, if auto_export is false) completes. Has no effect unless the export it depends on succeeds.",
            "default": false
          },
          "mode": {
            "type": "string",
            "const": "manual",
            "title": "Mode",
            "default": "manual"
          },
          "manual_config": {
            "$ref": "#/components/schemas/ManualTrainingConfig"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "dataset_id"
        ],
        "title": "ManualTrainingRequest",
        "examples": [
          {
            "base_model": "unsloth/Llama-3.2-3B-Instruct-bnb-4bit",
            "dataset_id": "00000000-0000-0000-0000-000000000010",
            "manual_config": {
              "gradient_accumulation_steps": 8,
              "learning_rate": 0.0002,
              "lora": {
                "alpha": 32,
                "dropout": 0.05,
                "r": 16,
                "target_modules": [
                  "q_proj",
                  "k_proj",
                  "v_proj",
                  "o_proj",
                  "gate_proj",
                  "up_proj",
                  "down_proj"
                ]
              },
              "num_train_epochs": 3,
              "optim": "adamw_8bit",
              "packing": false,
              "per_device_train_batch_size": 2
            },
            "mode": "manual",
            "project_id": "00000000-0000-0000-0000-000000000001",
            "training_name": "qa-policy-v1"
          }
        ]
      },
...
      "MlflowUrlResponse": {
        "properties": {
          "training_id": {
            "type": "string",
            "format": "uuid",
            "title": "Training Id"
          },
          "mlflow_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Run Id"
          },
          "mlflow_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "training_id",
          "mlflow_run_id",
          "mlflow_url"
        ],
        "title": "MlflowUrlResponse"
      },
...
      "Page_TrainingResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TrainingResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "maximum": 500.0,
            "exclusiveMinimum": 0.0,
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[TrainingResponse]"
      },
...
      "TrainingJobAcceptedResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "training_id": {
            "type": "string",
            "format": "uuid",
            "title": "Training Id"
          },
          "mlflow_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Run Id"
          },
          "mlflow_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Url"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "pending"
          },
          "websocket_url": {
            "type": "string",
            "title": "Websocket Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "training_id",
          "websocket_url"
        ],
        "title": "TrainingJobAcceptedResponse",
        "description": "202 Accepted body returned when a training job is enqueued."
      },
...
      "TrainingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "mode": {
            "$ref": "#/components/schemas/TrainingMode"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "celery_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Celery Task Id"
          },
          "base_model": {
            "type": "string",
            "title": "Base Model"
          },
          "training_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Name"
          },
          "mlflow_experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Experiment Id"
          },
          "mlflow_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Run Id"
          },
          "config_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config Json"
          },
          "best_metric_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Metric Value"
          },
          "best_params_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Params Json"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "auto_export": {
            "type": "boolean",
            "title": "Auto Export"
          },
          "auto_evaluate": {
            "type": "boolean",
            "title": "Auto Evaluate"
          },
          "auto_pipeline": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Pipeline"
          },
          "queue_state": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "processing",
                  "queued"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue State"
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position"
          },
          "owner_queue_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Queue Position"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "dataset_id",
          "mode",
          "status",
          "celery_task_id",
          "base_model",
          "training_name",
          "mlflow_experiment_id",
          "mlflow_run_id",
          "config_json",
          "best_metric_value",
          "best_params_json",
          "error_message",
          "started_at",
          "ended_at",
          "created_at",
          "updated_at",
          "auto_export",
          "auto_evaluate"
        ],
        "title": "TrainingResponse"
      },
...
    }
  },
...
}